app/mcp/requirements.txt
# Зависимости для MCP серверов
mcp==1.0.0
psycopg2-binary==2.9.9
PyYAML==6.0.1
asyncio